+++ MyProject/src/main/java/ru/bellintegrator/myproject/office/service/impl/OfficeServiceImpl.java
@@ -3,24 +3,19 @@ package ru.bellintegrator.myproject.office.service.impl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Scope;
-import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.bellintegrator.myproject.exceptions.OfficeServiceException;
 import ru.bellintegrator.myproject.office.dao.impl.OfficeDAOImpl;
 import ru.bellintegrator.myproject.office.model.Office;
 import ru.bellintegrator.myproject.office.service.OfficeService;
 import ru.bellintegrator.myproject.office.view.OfficeFilterView;
 import ru.bellintegrator.myproject.office.view.OfficeFilterViewList;
 import ru.bellintegrator.myproject.office.view.OfficeView;
-
 import java.util.ArrayList;
 import java.util.List;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 @Service
-//@Scope(proxyMode = ScopedProxyMode.INTERFACES)
 public class OfficeServiceImpl implements OfficeService {
 
     private final Logger logger = LoggerFactory.getLogger(OfficeServiceImpl.class);
@@ -58,14 +53,18 @@ public class OfficeServiceImpl implements OfficeService {
     @Override
     @Transactional
     public Office getOfficeById(Long id) {
+        Office office =DAO.getOfficeById(id);
+        if(office == null) throw new OfficeServiceException("Офиса с id " + id + " не существует");
         logger.info("Office get ID:" + id);
-        return DAO.getOfficeById(id);
+        return office;
     }
 
     @Override
     @Transactional
     public void update(OfficeView view) {
+        Long id = view.id;
         Office office = new Office(view.id, view.name, view.address, view.phone, view.isActive);
+        if(office == null) throw new OfficeServiceException("Организации с id " + id + " не существует");
         logger.info("Office update " + office.toString());
         DAO.update(office);
     }
@@ -82,6 +81,7 @@ public class OfficeServiceImpl implements OfficeService {
     @Transactional
     public void delete(Long id) {
         Office office = getOfficeById(id);
+        if(office == null) throw new OfficeServiceException("Офиса с id " + id + " не существует");
         logger.info("Office deleted ID:" + id);
         DAO.delete(office);
     }
